// src/ticket/quotes/quote-status.enum.ts
export enum QuoteStatus {
  CURRENT = 'CURRENT',
  PENDING_SUPERVISOR_APPROVAL = 'PENDING_SUPERVISOR_APPROVAL',
  SUPERVISOR_APPROVED = 'SUPERVISOR_APPROVED',
  SUPERVISOR_REJECTED = 'SUPERVISOR_REJECTED',
  SENT_TO_CLIENT = 'SENT_TO_CLIENT',
  AWAITING_CLIENT_RESPONSE = 'AWAITING_CLIENT_RESPONSE',
  CLIENT_APPROVED = 'CLIENT_APPROVED',
  CLIENT_REJECTED = 'CLIENT_REJECTED',
  ARCHIVED = 'ARCHIVED',
}